requirements.txt
click
pyedifice
pandas
scikit-learn
matplotlib
mlinsights
